move squid away from /packages
