+++ squid/README.md
