+++ squid/src/constants.ts
