+++ squid/src/main.ts
