+++ squid/src/model/index.ts
