+++ squid/src/types/events.ts
