+++ squid/src/types/v9190.ts
