+++ squid/src/types/v9220.ts
