+++ squid/src/types/v9250.ts
