+++ squid/src/types/v9300.ts
