+++ squid/src/types/v9370.ts
