+++ squid/src/types/v9380.ts
