+++ squid/src/types/v9420.ts
