+++ squid/src/util/JsonLog.ts
