+++ squid/src/util/getMultisigAddress.ts
@@ -1,5 +1,5 @@
 import { createKeyMulti, encodeAddress } from '@polkadot/util-crypto'
-import { env } from '../processor'
+import { env } from '../main'
 
 export const getMultisigAddress = (signatories: (string | Uint8Array)[], threshold: number) => {
   return encodeAddress(createKeyMulti(signatories, threshold), env.prefix)
