+++ squid/src/util/getProxyAccountId.ts
