+++ squid/src/util/index.ts
