+++ squid/typegen.json
